Move failure rate function to route folder
